requirements.txt
flask==3.1.0
azure-ai-projects>=1.0.0b7
azure-identity>=1.19.0
python-dotenv>=1.0.1
gunicorn>=22.0.0
